! added support for .gml output format ~ changed Color to byte for memory efficiency + updated .jar

+++ src/main/java/graph_gen_utils/chaco/ChacoWriterUnweighted.java
@@ -124,7 +124,7 @@ public class ChacoWriterUnweighted extends ChacoWriter {
 
 					String line = "0";
 					if (node.hasProperty("color")) {
-						Integer color = (Integer) node.getProperty("color");
+						Byte color = (Byte) node.getProperty("color");
 						line = color.toString();
 					}
